added Comment and Like Models

+++ post/services.py
@@ -1,6 +1,6 @@
 from . import models, schemas
 from core.db import get_db
-from fastapi import Depends
+from fastapi import Depends, HTTPException
 from sqlalchemy.orm import Session
 
 
@@ -13,14 +13,21 @@ def create_post(data:schemas.PostCreate, db:Session=Depends(get_db)):
 
 def get_posts(db:Session=Depends(get_db)):
     posts = db.query(models.Post).all()
+    if posts is None:
+        raise HTTPException(status_code=404, detail="No posts found")
     return posts
 
 def get_post_by_id(post_id:int, db:Session=Depends(get_db)):
     post = db.query(models.Post).filter(models.Post.id==post_id).first()
+    if post is None:
+        raise HTTPException(status_code=404, detail="Post not found")
     return post 
 
 def update_post(post_id:int, data:schemas.PostUpdate, db:Session=Depends(get_db)):
     post = db.query(models.Post).filter(models.Post.id==post_id).first()
+    if post is None:
+        raise HTTPException(status_code=404, detail="Post not found")
+    
     post.title = data.title
     post.content = data.content
     post.tags = data.tags
@@ -30,6 +37,8 @@ def update_post(post_id:int, data:schemas.PostUpdate, db:Session=Depends(get_db)
 
 def delete_post(post_id:int, db:Session=Depends(get_db)):
     post = db.query(models.Post).filter(models.Post.id==post_id).first()
+    if post is None:
+        raise HTTPException(status_code=404, detail="Post not found")
     db.delete(post)
     db.commit()
     return {"message":"Post deleted successfully"}
